Simplify integration test
