+++ pkg/integration/tests/test_list.go
@@ -160,8 +160,8 @@ var tests = []*components.IntegrationTest{
 	file.DiscardVariousChangesRangeSelect,
 	file.Gitignore,
 	file.RememberCommitMessageAfterFail,
+	file.StageChildrenRangeSelect,
 	file.StageRangeSelect,
-	file.StageRenamedRangeSelect,
 	filter_and_search.FilterCommitFiles,
 	filter_and_search.FilterFiles,
 	filter_and_search.FilterFuzzy,
